cleaning up the step definitions
